break up uplot-plugins
