+++ plugin/blocks/src/components/graph/components/lib/uplot-plugins/index.js
@@ -1,0 +1,9 @@
+/**
+ * Export all uPlot plugins from individual files
+ */
+
+export { legendLabelClickPlugin } from './legendLabelClickPlugin';
+export { isolateStackedPlugin } from './isolateStackedPlugin';
+export { preventLegendClickPlugin } from './preventLegendClickPlugin';
+export { seriesBarsPlugin } from './seriesBarsPlugin';
+export { tooltipPlugin } from './tooltipPlugin';
