+++ plugin/blocks/src/components/graph/components/lib/uplot-plugins/seriesBarsPlugin.js
@@ -3,217 +3,14 @@ import uPlot from 'uplot';
 /**
  * Local dependencies.
  */
-import Quadtree, { pointWithin } from './quadtree.js';
-import { distr, SPACE_BETWEEN } from './distr.js';
-
-/**
- * Plugin for handling legend label click events for site navigation
- *
- * @param {string} dimension - The dimension to check for clickable labels
- * @returns {Object} - The plugin object
- */
-export function legendLabelClickPlugin(dimension) {
-	return {
-		hooks: {
-			ready: [
-				(u) => {
-					// Function to handle legend label events
-					const handleLegendEvents = () => {
-						// Find all legend labels
-						const labels = u.root.querySelectorAll('.u-label');
-
-						// Add click event for specific dimensions
-						if ('atomic_site_id' === dimension) {
-							labels.forEach(label => {
-								// Skip labels that already have event handlers attached
-								if (label.dataset.hasClickEvent === 'true') {
-									return;
-								}
-
-								label.style.cursor = 'pointer';
-								const clickHandler = (e) => {
-									e.stopPropagation();
-									console.log(`Clicked label: ${label.textContent}`);
-
-									var site_url = label.textContent;
-									if (isNaN(site_url)) {
-										site_url = site_url.replace(/\./g, "-");
-									}
-									window.location.href = `/sites/${site_url}/metrics/`;
-								};
-
-								label.addEventListener('click', clickHandler);
-
-								// Mark this label as having click event handler attached
-								label.dataset.hasClickEvent = 'true';
-							});
-						}
-					};
-
-					// Initial setup with a delay to ensure DOM is ready
-					setTimeout(handleLegendEvents, 100);
-
-					// Set up a MutationObserver to watch for changes to the legend
-					const observer = new MutationObserver((mutations) => {
-						// Re-apply event handlers when the DOM changes
-						handleLegendEvents();
-					});
-
-					// Start observing the chart root for changes
-					observer.observe(u.root, {
-						childList: true,
-						subtree: true
-					});
-
-					// Store the observer for cleanup
-					u.legendObserver = observer;
-				}
-			],
-			destroy: [
-				(u) => {
-					// Clean up the observer when the chart is destroyed
-					if (u.legendObserver) {
-						u.legendObserver.disconnect();
-					}
-
-					// Remove data attributes from labels
-					const labels = u.root.querySelectorAll('.u-label');
-					labels.forEach(label => {
-						if (label.dataset.hasClickEvent === 'true') {
-							delete label.dataset.hasClickEvent;
-						}
-					});
-				}
-			]
-		}
-	};
-}
-
-/**
- * Plugin that customizes legend clicks
- * When a legend item is clicked, it shows only that series and hides all others
- */
-export function isolateStackedPlugin(originalData) {
-	let activeSeriesIdx = null; // Track which series is currently active
-	let stackedData = null; // Store the stacked data
-
-	return {
-		hooks: {
-			ready: (u) => {
-
-				// Store the stacked data that's currently in the chart
-				if (u.data && Array.isArray(u.data) && u.data.length > 0) {
-					stackedData = [...u.data];
-				}
-
-				// Find the legend element
-				const legendEl = u.root.querySelector(".u-legend");
-
-				if (legendEl) {
-					// Add capture phase event listener to intercept clicks before they reach uPlot's handlers
-					legendEl.addEventListener("click", (e) => {
-						// Stop propagation and prevent default to block uPlot's default behavior
-						e.stopPropagation();
-						e.preventDefault();
-
-						// Find the closest legend item (tr element)
-						const target = e.target;
-						const legendItem = target.closest("tr");
-
-						if (!legendItem) {
-							return false;
-						}
-
-						// Find all legend items (tr elements)
-						const legendItems = legendEl.querySelectorAll("tr");
-
-						// Get the index of the clicked legend item
-						let clickedIdx = Array.from(legendItems).indexOf(legendItem);
-
-						// If the same item is clicked again, show all series
-						if (activeSeriesIdx === clickedIdx) {
-							console.log("Showing all series");
-							// Remove u-off class from all legend items
-							legendItems.forEach(item => {
-								item.classList.remove("u-off");
-							});
-							activeSeriesIdx = null;
-
-							// Restore the stacked data
-							if (stackedData) {
-								u.setData(stackedData);
-							}
-						} else {
-							console.log("Showing only series", clickedIdx);
-							// Apply u-off class to all legend items except the clicked one
-							legendItems.forEach((item, idx) => {
-								if (idx !== clickedIdx) {
-									item.classList.add("u-off");
-								} else {
-									item.classList.remove("u-off");
-								}
-							});
-							activeSeriesIdx = clickedIdx;
-
-							// Replace the data with original data for the clicked series
-							if (originalData && Array.isArray(originalData) && originalData.length > clickedIdx) {
-								++clickedIdx; // Adjust for the x-axis series
-
-								// Create a new data array with just the x-axis values and the clicked series
-								const newData = stackedData.map((series, idx) => {
-									if (idx === 0) {
-										// Keep x-axis values
-										return series;
-									} else if (idx === clickedIdx) {
-										// Use the original unstacked data for this series
-										return originalData[idx];
-									} else {
-										// For other series, create an array of the same length as the x-axis but with null values
-										return Array(stackedData[0].length).fill(null);
-									}
-								});
-
-								// Update the graph with the new data
-								u.setData(newData);
-							}
-						}
-
-						return false;
-					}, true); // true for capture phase
-				} else {
-					console.warn("Legend element not found");
-				}
-			}
-		}
-	};
-}
-
-/**
- * Plugin that prevents legend clicks for non-stacked bar charts
- */
-export function preventLegendClickPlugin() {
-	return {
-		hooks: {
-			ready: (u) => {
-				// Find the legend element
-				const legendEl = u.root.querySelector(".u-legend");
-
-				if (legendEl) {
-					// Add capture phase event listener to intercept clicks before they reach uPlot's handlers
-					legendEl.addEventListener("click", (e) => {
-						// Stop propagation and prevent default to block uPlot's default behavior
-						e.stopPropagation();
-						e.preventDefault();
-						return false;
-					}, true); // true for capture phase
-				}
-			}
-		}
-	};
-}
+import Quadtree, { pointWithin } from '../quadtree.js';
+import { distr, SPACE_BETWEEN } from '../distr.js';
 
 /**
  * This from https://github.com/leeoniya/uPlot/blob/4315544c319a2c9d561ebd29f54d06a034fb16f6/demos/grouped-bars.js
+ *
+ * @param {Object} opts - Options for the plugin
+ * @returns {Object} - The plugin object
  */
 export function seriesBarsPlugin(opts = {}) {
 	let pxRatio;
