+++ plugin/blocks/src/components/graph/components/lib/uplot-plugins/tooltipPlugin.js
@@ -1,0 +1,13 @@
+/**
+ * Plugin for adding tooltips to uPlot charts
+ *
+ * @returns {Object} - The plugin object
+ */
+export function tooltipPlugin() {
+	return {
+		hooks: {
+			// Add tooltip functionality here if needed
+			// This plugin was referenced in the imports but not implemented in the original file
+		}
+	};
+}
